+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added [CLIPScore](models/hessel2021/Readme.md)
+- Added a [QA SRL Parser](models/fitzgerald2018/Readme.md)
 
 ## [v0.1.2](https://github.com/danieldeutsch/repro/releases/tag/v0.1.1) - 2021-10-07
 ### Changed