+++ Changelog.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [BLANC](models/vasilyev2020/Readme.md)
 - Added [METEOR](models/denkowski2014/Readme.md)
 
+### Fixed
+- Fixed an error in [Lite3Pyramid](models/zhang2021/Readme.md) by updating to a newer version of the code.
+
 ## [v0.1.2](https://github.com/danieldeutsch/repro/releases/tag/v0.1.1) - 2021-10-07
 ### Changed
 - Changed backend of Lite3Pyramid to use our own fork of the official repo with some modifications.