+++ Changelog.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [METEOR](models/denkowski2014/Readme.md)
 - Added a role question generator from [Pyatkin et al. (2021)](models/pyatkin2021/Readme.md)
 
+### Fixed
+- Fixed an error in [Lite3Pyramid](models/zhang2021/Readme.md) by updating to a newer version of the code.
+
 ## [v0.1.2](https://github.com/danieldeutsch/repro/releases/tag/v0.1.1) - 2021-10-07
 ### Changed
 - Changed backend of Lite3Pyramid to use our own fork of the official repo with some modifications.